[CORE-1581] feat: make upserting wallet optional (#106)
# Description
By default, we have behavior when sdk upserts a wallet after creating
the token. This is the bottleneck in dial-app while we are creating a
new user because the upsert wallet endpoint (`api/v1/wallets/me`)
invokes with `x-client-name: dialect-sdk`. For example, when we want to
handle created user from the dial-app we won't be able to do that,
because the user will be created according to `dialect-sdk` HTTP headers
context in a request